+++ lib/asset/sql.go
@@ -132,3 +132,49 @@ func (k *EventKind) Scan(value interface{}) error {
 
 	return nil
 }
+
+// Value implements the driver.Valuer interface.
+// Simply returns the string representation of the FunctionStatus.
+func (fs *FunctionStatus) Value() (driver.Value, error) {
+	return fs.String(), nil
+}
+
+// Scan implements the sql.Scanner interface.
+// Simply decodes a string into the FunctionStatus.
+func (fs *FunctionStatus) Scan(value interface{}) error {
+	s, ok := value.(string)
+	if !ok {
+		return errors.NewInternal("cannot scan function status: invalid string")
+	}
+
+	v, ok := FunctionStatus_value[s]
+	if !ok {
+		return errors.NewInternal("cannot scan function status: unknown value")
+	}
+	*fs = FunctionStatus(v)
+
+	return nil
+}
+
+// Value implements the driver.Valuer interface.
+// Simply returns the string representation of the FunctionStatus.
+func (fak *FailedAssetKind) Value() (driver.Value, error) {
+	return fak.String(), nil
+}
+
+// Scan implements the sql.Scanner interface.
+// Simply decodes a string into the FunctionStatus.
+func (fak *FailedAssetKind) Scan(value interface{}) error {
+	s, ok := value.(string)
+	if !ok {
+		return errors.NewInternal("cannot scan failed asset kind: invalid string")
+	}
+
+	v, ok := FailedAssetKind_value[s]
+	if !ok {
+		return errors.NewInternal("cannot scan failed asset kind: unknown value")
+	}
+	*fak = FailedAssetKind(v)
+
+	return nil
+}